Fix crashing on vertex click after deselect
This was a nasty long-timer bug. A lot of unfortunate crashes before autosave was there. It
usually happened when one was trying to click on the selected surface vertex, failed and
then tried again. The selection widget joints were still hitTested, but as there was no
selection anymore, the app crashed while trying to access a vertex of a non-existing
surface.