Fix texture coords not matching tex crop area
After hitting TAB and changing to nex texture in texture mode, joints were not matchin the
actual texture coordinates. Fix here. There is another problem though. There is a problem
with videos that have not been allocated yet. There is no size for the texture, thus it is
not possible calculate the texture coods.