added 4 more FBO sources
Added CircleSource, QuadSource, BrickSource and SyphonSource.

The SyphonFBO is disabled since it would not run on rpi.

[OSX only]
For Syphon to work you need to install ofxSyphon addon and add "ofxSyphon"
to the addons.make of ofxPiMapper. Then uncomment all the syphon lines in the code. So far the SyphonClient will choose the first Syphon source presented to him. ( You might want to start the server later ).